Reorganizing IDE entrypoint
